Ported enhancements and fixes from https://github.com/hazelcast/hazelcast/pull/12930 (Refactoring client listener service) and https://github.com/hazelcast/hazelcast/pull/12765 (Close connections when heartbeat timeout).
