+++ hazelcast/include/hazelcast/client/spi/impl/AbstractClientInvocationService.h
@@ -124,8 +124,6 @@ namespace hazelcast {
 
                     void registerInvocation(const boost::shared_ptr<ClientInvocation> &clientInvocation);
 
-                    bool isAllowedToSendRequest(connection::Connection &connection, ClientInvocation &invocation);
-
                     bool writeToConnection(connection::Connection &connection,
                                            const boost::shared_ptr<protocol::ClientMessage> &clientMessage);
