+++ hazelcast/src/hazelcast/client/spi/impl/listener/SmartClientListenerService.cpp
@@ -20,11 +20,9 @@
 #include "hazelcast/util/UuidUtil.h"
 #include "hazelcast/client/Member.h"
 #include "hazelcast/client/spi/impl/listener/SmartClientListenerService.h"
-#include "hazelcast/client/spi/ClientContext.h"
 #include "hazelcast/client/spi/ClientClusterService.h"
 #include "hazelcast/client/spi/LifecycleService.h"
 #include "hazelcast/client/connection/ClientConnectionManagerImpl.h"
-#include "hazelcast/client/connection/Connection.h"
 #include "hazelcast/client/spi/impl/ListenerMessageCodec.h"
 
 namespace hazelcast {
@@ -35,11 +33,17 @@ namespace hazelcast {
                     SmartClientListenerService::SmartClientListenerService(ClientContext &clientContext,
                                                                            int32_t eventThreadCount,
                                                                            int32_t eventQueueCapacity)
-                            : AbstractClientListenerService(clientContext, eventThreadCount, eventQueueCapacity),
-                              clientConnectionManager(clientContext.getConnectionManager()) {
-                        AbstractClientInvocationService &invocationService = (AbstractClientInvocationService &) clientContext.getInvocationService();
-                        invocationTimeoutMillis = invocationService.getInvocationTimeoutMillis();
-                        invocationRetryPauseMillis = invocationService.getInvocationRetryPauseMillis();
+                            : AbstractClientListenerService(clientContext, eventThreadCount, eventQueueCapacity) {
+                    }
+
+
+                    void SmartClientListenerService::start() {
+                        AbstractClientListenerService::start();
+
+                        registrationExecutor.scheduleAtFixedRate(
+                                boost::shared_ptr<util::Runnable>(new AsyncConnectToAllMembersTask(
+                                        boost::static_pointer_cast<SmartClientListenerService>(shared_from_this()))),
+                                1000, 1000);
                     }
 
                     std::string
@@ -52,24 +56,7 @@ namespace hazelcast {
 */
                         trySyncConnectToAllMembers();
 
-                        boost::shared_ptr<util::Callable<std::string> > task(
-                                new RegisterListenerTask("SmartClientListenerService::registerListener",
-                                                         shared_from_this(), listenerMessageCodec, handler));
-                        return registrationExecutor.submit<std::string>(task)->get();
-                    }
-
-                    bool SmartClientListenerService::deregisterListener(const std::string &registrationId) {
-                        //This method should not be called from registrationExecutor
-/*                      TODO
-                        assert (!Thread.currentThread().getName().contains("eventRegistration"));
-*/
-
-                        boost::shared_ptr<util::Future<bool> > future = registrationExecutor.submit(
-                                boost::shared_ptr<util::Callable<bool> >(
-                                        new DeregisterListenerTask("SmartClientListenerService::deregisterListener",
-                                                                   shared_from_this(), registrationId)));
-
-                        return future->get();
+                        return AbstractClientListenerService::registerListener(listenerMessageCodec, handler);
                     }
 
                     void SmartClientListenerService::trySyncConnectToAllMembers() {
@@ -136,233 +123,21 @@ namespace hazelcast {
                         util::sleepmillis(invocationRetryPauseMillis);
                     }
 
-                    void
-                    SmartClientListenerService::invoke(const ClientRegistrationKey &registrationKey,
-                                                       const boost::shared_ptr<connection::Connection> &connection) {
-                        //This method should only be called from registrationExecutor
-/*                      TODO
-                        assert (Thread.currentThread().getName().contains("eventRegistration"));
-*/
-
-                        boost::shared_ptr<ConnectionRegistrationsMap> registrationMap = registrations.get(
-                                registrationKey);
-                        if (registrationMap->find(connection) != registrationMap->end()) {
-                            return;
-                        }
-
-                        const boost::shared_ptr<ListenerMessageCodec> &codec = registrationKey.getCodec();
-                        std::auto_ptr<protocol::ClientMessage> request = codec->encodeAddRequest(true);
-                        boost::shared_ptr<EventHandler<protocol::ClientMessage> > handler = registrationKey.getHandler();
-                        handler->beforeListenerRegister();
-
-                        boost::shared_ptr<ClientInvocation> invocation = ClientInvocation::create(clientContext,
-                                                                                                  request, "",
-                                                                                                  connection);
-                        invocation->setEventHandler(handler);
-
-                        boost::shared_ptr<protocol::ClientMessage> clientMessage = invocation->invokeUrgent()->get();
-
-                        std::string serverRegistrationId = codec->decodeAddResponse(*clientMessage);
-                        handler->onListenerRegister();
-                        int64_t correlationId = invocation->getClientMessage()->getCorrelationId();
-                        ClientEventRegistration registration(serverRegistrationId, correlationId, connection, codec);
-
-                        (*registrationMap)[connection] = registration;
+                    bool SmartClientListenerService::registersLocalOnly() const {
+                        return true;
                     }
 
-                    void SmartClientListenerService::connectionAdded(
-                            const boost::shared_ptr<connection::Connection> &connection) {
-                        //This method should only be called from registrationExecutor
-/*                      TODO
-                        assert (Thread.currentThread().getName().contains("eventRegistration"));
-*/
-                        registrationExecutor.execute(
-                                boost::shared_ptr<util::Runnable>(
-                                        new ConnectionAddedTask("SmartClientListenerService::connectionAdded",
-                                                                shared_from_this(), connection)));
-                    }
-
-                    void SmartClientListenerService::connectionRemoved(
-                            const boost::shared_ptr<connection::Connection> &connection) {
-                        //This method should only be called from registrationExecutor
-/*                      TODO
-                        assert (Thread.currentThread().getName().contains("eventRegistration"));
-*/
-
-                        registrationExecutor.execute(
-                                boost::shared_ptr<util::Runnable>(
-                                        new ConnectionRemovedTask("SmartClientListenerService::connectionRemoved",
-                                                                  shared_from_this(), connection)));
-                    }
-
-                    void SmartClientListenerService::heartbeatResumed(
-                            const boost::shared_ptr<connection::Connection> &connection) {
-                        //This method should only be called from registrationExecutor
-/*                      TODO
-                        assert (Thread.currentThread().getName().contains("eventRegistration"));
-*/
-                        registrationExecutor.execute(
-                                boost::shared_ptr<util::Runnable>(new HearbeatResumedTask(*this, connection)));
-                    }
-
-                    void SmartClientListenerService::heartbeatStopped(
-                            const boost::shared_ptr<connection::Connection> &connection) {
-                        //no op
-                    }
-
-                    void
-                    SmartClientListenerService::invokeFromInternalThread(const ClientRegistrationKey &registrationKey,
-                                                                         const boost::shared_ptr<connection::Connection> &connection) {
-                        //This method should only be called from registrationExecutor
-/*                      TODO
-                        assert (Thread.currentThread().getName().contains("eventRegistration"));
-*/
-                        try {
-                            invoke(registrationKey, connection);
-                        } catch (exception::IOException &) {
-                            boost::shared_ptr<std::vector<ClientRegistrationKey> > failedRegsToConnection = failedRegistrations.get(
-                                    connection);
-                            if (failedRegsToConnection.get() == NULL) {
-                                failedRegsToConnection = boost::shared_ptr<std::vector<ClientRegistrationKey> >(
-                                        new std::vector<ClientRegistrationKey>());
-                                failedRegistrations.put(connection, failedRegsToConnection);
-                            }
-                            failedRegsToConnection->push_back(registrationKey);
-                        } catch (exception::IException &e) {
-                            logger.warning() << "Listener " << registrationKey
-                                             << " can not be added to a new connection: "
-                                             << *connection << ", reason: " << e.getMessage();
-                        }
-                    }
-
-                    void SmartClientListenerService::start() {
-                        clientConnectionManager.addConnectionListener(shared_from_this());
-                        clientConnectionManager.addConnectionHeartbeatListener(shared_from_this());
-
-                        registrationExecutor.scheduleAtFixedRate(
-                                boost::shared_ptr<util::Runnable>(new AsyncConnectToAllMembersTask(shared_from_this())),
-                                1000, 1000);
-                    }
-
-                    std::string SmartClientListenerService::registerListenerInternal(
-                            const boost::shared_ptr<ListenerMessageCodec> &listenerMessageCodec,
-                            const boost::shared_ptr<EventHandler<protocol::ClientMessage> > &handler) {
-                        std::string userRegistrationId = util::UuidUtil::newUnsecureUuidString();
-
-                        ClientRegistrationKey registrationKey(userRegistrationId, handler, listenerMessageCodec);
-                        registrations.put(registrationKey, boost::shared_ptr<ConnectionRegistrationsMap>(
-                                new ConnectionRegistrationsMap()));
-                        BOOST_FOREACH (const boost::shared_ptr<connection::Connection> &connection,
-                                       clientConnectionManager.getActiveConnections()) {
+                    void SmartClientListenerService::asyncConnectToAllMembersInternal() {
+                        std::vector<Member> memberList = clientContext.getClientClusterService().getMemberList();
+                        BOOST_FOREACH (const Member &member, memberList) {
                                         try {
-                                            invoke(registrationKey, connection);
-                                        } catch (exception::IException &e) {
-                                            if (connection->isAlive()) {
-                                                deregisterListenerInternal(userRegistrationId);
-                                                throw (exception::ExceptionBuilder<exception::HazelcastException>(
-                                                        "SmartClientListenerService::RegisterListenerTask::call")
-                                                        << "Listener can not be added " << e).build();
-                                            }
+                                            clientContext.getConnectionManager().getOrTriggerConnect(
+                                                    member.getAddress());
+                                        } catch (exception::IOException &) {
+                                            return;
                                         }
                                     }
-                        return userRegistrationId;
-                    }
 
-                    bool SmartClientListenerService::deregisterListenerInternal(const std::string &userRegistrationId) {
-                        //This method should only be called from registrationExecutor
-/*                      TODO
-                        assert (Thread.currentThread().getName().contains("eventRegistration"));
-*/
-                        ClientRegistrationKey key(userRegistrationId);
-                        boost::shared_ptr<ConnectionRegistrationsMap> registrationMap = registrations.get(key);
-                        if (registrationMap.get() == NULL) {
-                            return false;
-                        }
-                        bool successful = true;
-
-                        for (ConnectionRegistrationsMap::iterator it = registrationMap->begin();
-                             it != registrationMap->end();) {
-                            ClientEventRegistration &registration = (*it).second;
-                            boost::shared_ptr<connection::Connection> subscriber = registration.getSubscriber();
-                            try {
-                                const boost::shared_ptr<ListenerMessageCodec> &listenerMessageCodec = registration.getCodec();
-                                const std::string &serverRegistrationId = registration.getServerRegistrationId();
-                                std::auto_ptr<protocol::ClientMessage> request = listenerMessageCodec->encodeRemoveRequest(
-                                        serverRegistrationId);
-                                boost::shared_ptr<ClientInvocation> invocation = ClientInvocation::create(clientContext,
-                                                                                                          request, "",
-                                                                                                          subscriber);
-                                invocation->invoke()->get();
-                                removeEventHandler(registration.getCallId());
-
-                                ConnectionRegistrationsMap::iterator oldEntry = it;
-                                ++it;
-                                registrationMap->erase(oldEntry);
-                            } catch (exception::IException &e) {
-                                ++it;
-
-                                if (subscriber->isAlive()) {
-                                    successful = false;
-                                    std::ostringstream endpoint;
-                                    if (subscriber->getRemoteEndpoint().get()) {
-                                        endpoint << *subscriber->getRemoteEndpoint();
-                                    } else {
-                                        endpoint << "null";
-                                    }
-                                    logger.warning() << "SmartClientListenerService::deregisterListenerInternal"
-                                                     << "Deregistration of listener with ID " << userRegistrationId
-                                                     << " has failed to address " << subscriber->getRemoteEndpoint()
-                                                     << e;
-                                }
-                            }
-                        }
-                        if (successful) {
-                            registrations.remove(key);
-                        }
-                        return successful;
-                    }
-
-                    void SmartClientListenerService::connectionAddedInternal(
-                            const boost::shared_ptr<connection::Connection> &connection) {
-                        BOOST_FOREACH(const ClientRegistrationKey &registrationKey, registrations.keys()) {
-                                        invokeFromInternalThread(registrationKey, connection);
-                                    }
-                    }
-
-                    void SmartClientListenerService::connectionRemovedInternal(
-                            const boost::shared_ptr<connection::Connection> &connection) {
-                        failedRegistrations.remove(connection);
-                        typedef std::vector<std::pair<ClientRegistrationKey, boost::shared_ptr<ConnectionRegistrationsMap> > > ENTRY_VECTOR;
-                        BOOST_FOREACH(const ENTRY_VECTOR::value_type &registrationMapEntry, registrations.entrySet()) {
-                                        boost::shared_ptr<ConnectionRegistrationsMap> registrationMap = registrationMapEntry.second;
-                                        ConnectionRegistrationsMap::iterator foundRegistration = registrationMap->find(
-                                                connection);
-                                        if (foundRegistration != registrationMap->end()) {
-                                            removeEventHandler(foundRegistration->second.getCallId());
-                                            registrationMap->erase(foundRegistration);
-                                            registrations.put(registrationMapEntry.first,
-                                                              registrationMap);
-                                        }
-                                    }
-                    }
-
-                    SmartClientListenerService::HearbeatResumedTask::HearbeatResumedTask(
-                            SmartClientListenerService &listenerService,
-                            const boost::shared_ptr<connection::Connection> &connection) : listenerService(
-                            listenerService.shared_from_this()), connection(connection) {}
-
-                    const std::string SmartClientListenerService::HearbeatResumedTask::getName() const {
-                        return "SmartClientListenerService::HearbeatResumedTask";
-                    }
-
-                    void SmartClientListenerService::HearbeatResumedTask::run() {
-                        boost::shared_ptr<std::vector<ClientRegistrationKey> > registrationKeys = listenerService->failedRegistrations.get(
-                                connection);
-                        if (registrationKeys.get()) {
-                            BOOST_FOREACH(const ClientRegistrationKey &registrationKey, *registrationKeys) {
-                                            listenerService->invokeFromInternalThread(registrationKey, connection);
-                                        }
-                        }
                     }
 
                     SmartClientListenerService::AsyncConnectToAllMembersTask::AsyncConnectToAllMembersTask(
@@ -370,51 +145,11 @@ namespace hazelcast {
                             listenerService) {}
 
                     void SmartClientListenerService::AsyncConnectToAllMembersTask::run() {
-                        std::vector<Member> memberList = listenerService->clientContext.getClientClusterService().getMemberList();
-                        BOOST_FOREACH (const Member &member, memberList) {
-                                        try {
-                                            listenerService->clientContext.getConnectionManager().getOrTriggerConnect(
-                                                    member.getAddress());
-                                        } catch (exception::IOException &) {
-                                            return;
-                                        }
-                                    }
+                        listenerService->asyncConnectToAllMembersInternal();
                     }
 
                     const std::string SmartClientListenerService::AsyncConnectToAllMembersTask::getName() const {
                         return "SmartClientListenerService::AsyncConnectToAllMembersTask";
-                    }
-
-                    SmartClientListenerService::ConnectionRemovedTask::ConnectionRemovedTask(
-                            const std::string &taskName,
-                            const boost::shared_ptr<SmartClientListenerService> &listenerService,
-                            const boost::shared_ptr<connection::Connection> &connection) : taskName(taskName),
-                                                                                           listenerService(
-                                                                                                   listenerService),
-                                                                                           connection(connection) {}
-
-                    const std::string  SmartClientListenerService::ConnectionRemovedTask::getName() const {
-                        return taskName;
-                    }
-
-                    void  SmartClientListenerService::ConnectionRemovedTask::run() {
-                        listenerService->connectionRemovedInternal(connection);
-                    }
-
-                    bool SmartClientListenerService::ConnectionPointerLessComparator::operator()(
-                            const boost::shared_ptr<connection::Connection> &lhs,
-                            const boost::shared_ptr<connection::Connection> &rhs) const {
-                        if (lhs == rhs) {
-                            return false;
-                        }
-                        if (!lhs.get()) {
-                            return true;
-                        }
-                        if (!rhs.get()) {
-                            return false;
-                        }
-
-                        return *lhs < *rhs;
                     }
                 }
             }
